v1.0

+++ microservice-provider-user/src/main/java/com/cloud/web/controller/UserController.java
@@ -29,7 +29,7 @@ public class UserController {
      */
     @GetMapping("/{id}")
     public User findById(@PathVariable Long id) {
-        User findOne = this.userMapper.selectByPrimaryKey(id);
+        User findOne = this.userMapper.selectByPrimaryKey(id.intValue());
         return findOne;
     }
